chore: separate fetching logic from layout component
