+++ components/Launches.tsx
@@ -1,5 +1,6 @@
 import { useEffect, useState } from 'react';
 import { Launch } from '../types/Launch';
+import { fetchLaunches } from './fetchLaunches';
 
 export function Launches() {
   const [launches, setLaunches] = useState<Launch[]>([]);
@@ -32,32 +33,4 @@ export function Launches() {
       ))}
     </article>
   );
-}
-
-function fetchLaunches(): Promise<Launch[]> {
-  const headers = new Headers({ 'Content-Type': 'application/json' });
-
-  const body = JSON.stringify({
-    query: {},
-    options: {
-      offset: 0,
-      limit: 10,
-      page: 1,
-      select: ['id', 'name', 'date_utc', 'cores.core', 'payloads', 'links.patch.small', 'success', 'failures.reason'],
-      populate: [
-        {
-          path: 'payloads',
-          select: ['id', 'type'],
-        },
-      ],
-    },
-  });
-
-  return fetch('https://api.spacexdata.com/v5/launches/query', {
-    method: 'POST',
-    headers,
-    body,
-  })
-    .then((response) => response.json())
-    .then((json) => json.docs);
 }
